+++ crates/sevenmark-server/src/api/v0/routes/openapi.rs
@@ -1,4 +1,4 @@
-use super::parse::openapi::ParseApiDoc;
+use super::render::openapi::RenderApiDoc;
 use utoipa::OpenApi;
 
 #[derive(OpenApi)]
@@ -8,7 +8,7 @@ pub struct V0ApiDoc;
 impl V0ApiDoc {
     pub fn merged() -> utoipa::openapi::OpenApi {
         let mut openapi = Self::openapi();
-        openapi.merge(ParseApiDoc::openapi());
+        openapi.merge(RenderApiDoc::openapi());
         openapi
     }
 }